requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
bittensor>=7.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
